% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_diff_cell_marker_net.R
\name{combine_diff_cell_marker_net}
\alias{combine_diff_cell_marker_net}
\title{Build a cell-cell network from a custom cell-type marker table}
\usage{
combine_diff_cell_marker_net(expression_data, geneList = NULL, cell_markers)
}
\arguments{
\item{expression_data}{Expression matrix, log(CPM/TPM), with gene symbols as row names.}

\item{geneList}{A character vector of gene symbols (e.g. differentially expressed genes). Default is NULL.}

\item{cell_markers}{A marker table with columns 'cellName' and 'geneID'.}
}
\value{
A list containing the following objects:
  \item{enricher}{Over-representation result on the cell-type markers.}
  \item{marker}{The cell-type marker table used.}
  \item{bnObject}{Deconvolution cell link from Bayesian Network.}
}
\description{
Build a cell-cell network from a custom cell-type marker table
}
