% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype_compareEnrichment.r
\name{compareEnrichment}
\alias{compareEnrichment}
\title{Compare functional enrichment across deconvolution cell types}
\usage{
compareEnrichment(
  DCL_Object = DCL_Object,
  celltypes,
  enrichType,
  showCategory = 5,
  pvalueCutoff = 1,
  pAdjustMethod = "BH",
  minGSSize = 1,
  qvalueCutoff = 1
)
}
\arguments{
\item{DCL_Object}{The object returned by DCL_net or DCL_GSEA_net.}

\item{celltypes}{Cell types to compare, matched by name. E.g. "Monocyte".}

\item{enrichType}{One of 'GO', 'Reactome', 'KEGG'.}

\item{showCategory}{Number of categories to show per cell type. Default 5.}

\item{pvalueCutoff}{P-value cutoff for enrichment. Default 1.}

\item{pAdjustMethod}{Multiple-testing adjustment method. Default "BH".}

\item{minGSSize}{Minimum gene set size. Default 1.}

\item{qvalueCutoff}{Q-value cutoff for enrichment. Default 1.}
}
\value{
A list containing the following objects:
  \item{enrichmentResult}{compareCluster result across the chosen cell types.}
  \item{comparePlot}{Dot plot of the compared enrichment.}
}
\description{
Compare functional enrichment across deconvolution cell types
}
