% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCL_LR_plot.R
\name{DCL_GSEA_net_interaction}
\alias{DCL_GSEA_net_interaction}
\title{Calculate LR in Deconvolution Cell Link}
\usage{
DCL_GSEA_net_interaction(DCL_Object = DCL_Object, deg = NULL)
}
\arguments{
\item{DCL_Object}{Deconvolution Cell Link object from DCL_GSEA or DCL_net.}

\item{deg}{A character of differentially expressed genes (DEgenes), Default is NULL.}
}
\value{
A list containing the following objects:
  \item{results}{L-R results.}
  \item{LR_plot}{L-R plot.}
}
\description{
Calculate LR in Deconvolution Cell Link
}
