% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCL_LR_plot.R
\name{DCL_LR_plot}
\alias{DCL_LR_plot}
\title{Calculate LR in Deconvolution Cell Link}
\usage{
DCL_LR_plot(
  DCL_Object = DCL_Object,
  deg = NULL,
  expression_data = NULL,
  cor_method = "pearson",
  cor_threshold = 0
)
}
\arguments{
\item{DCL_Object}{Deconvolution Cell Link object from DCL_GSEA or DCL_net.}

\item{deg}{A character of differentially expressed genes (DEgenes), Default is NULL.}

\item{expression_data}{data frame of expression data}

\item{cor_method}{correlation method of ligand and receptor, one of "pearson", "kendall" and "spearman"}

\item{cor_threshold}{the threshold of correlation}
}
\value{
A list containing the following objects:
  \item{results}{L-R results.}
  \item{LR_plot}{L-R plot.}
}
\description{
Calculate LR in Deconvolution Cell Link
}
