% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DCL_net.R
\name{.onLoad}
\alias{.onLoad}
\title{Calculate Deconvolution Cell Link for Bulk RNAseq data}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{expression_data}{A data frame containing the expression data (CPM/TPM).}

\item{geneList}{A gene list, such as a list of differentially expressed genes (DEgenes), to be used for SSMD calculation. Default is NULL.}

\item{tissueType}{The tissue type of the expression data. Choose one of the following: 'Inflammatory', 'Central Nervous System', 'Hematopoietic System', 'Blood'.}
}
\value{
A list containing the following objects:
  \item{bnObject}{Deconvolution cell link from Bayesian Network.}
  \item{cells_proportion}{Deconvolution Cell proportion from SSMD.}
}
\description{
This package provides functionality for [brief description of what your package does].

To install the required packages, please make sure you have BiocManager and remotes installed. If not, you can install them using the following code:
\preformatted{
if (!require(BiocManager))
  install.packages("BiocManager")

if (!require(remotes))
  install.packages("remotes")
}

To install this package, you can use the following code:
\preformatted{
BiocManager::install("JH-42/DevconCellLink")
}
}
\references{
- Jun-Hao Wang#, Zhao-Qian Zhong#, Hai-Hua Luo#, Qi-Zheng Han, Kan Wu, Li Chen, Yong Jiang*. Comparative Analysis of Brain Inflammation Reveals New Insights into Sepsis Associated Encephalopathy Progression
}
