% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LRI_mouse}
\alias{LRI_mouse}
\title{LRI mouse data}
\format{
A list with elements: ...
}
\source{
...
}
\usage{
LRI_mouse
}
\description{
A dataset containing LRI mouse data.
}
\keyword{datasets}
