% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_diff_cell_marker_net.R
\name{combine_diff_cell_marker_net}
\alias{combine_diff_cell_marker_net}
\title{Title}
\usage{
combine_diff_cell_marker_net(expression_data, geneList = NULL, cell_markers)
}
\arguments{
\item{expression_data}{the expression_data input: CPM/TPM}

\item{geneList}{DEgene list or other gene list， defult=NULL}

\item{tissueType}{the one of 'Inflammatory', 'Central Nervous System', 'Hematopoietic System','Blood'}
}
\value{
A list containing the following objects:
  \item{enrichmentResult}{Enrichment analysis of cell types specific marker.}
}
\description{
Title
}
