% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype_compareEnrichment.r
\name{compareEnrichment}
\alias{compareEnrichment}
\title{Title}
\usage{
compareEnrichment(
  celltypeObject,
  celltypes,
  enrichType,
  showCategory = 5,
  pvalueCutoff = 1,
  pAdjustMethod = "BH",
  minGSSize = 1,
  qvalueCutoff = 1
)
}
\arguments{
\item{celltypeObject}{the object from celltype or celltype_combine}

\item{celltypes}{choose your favorite cell types E.g: "Monocyte"}

\item{enrichType}{the one of 'GO', 'Reactome', 'KEGG'}
}
\value{
A list containing the following objects:
  \item{bnObject}{Deconvolution cell link from Bayesian Network.}
  \item{cells_proportion}{Deconvolution Cell proportion from SSMD.}
}
\description{
Title
}
